feat(web): tenant rbac manager pages

+++ src/main/kotlin/com/lovelycatv/crystalframework/tenant/controller/manager/role/ManagerTenantRoleController.kt
@@ -2,16 +2,19 @@ package com.lovelycatv.crystalframework.tenant.controller.manager.role
 
 import com.lovelycatv.crystalframework.rbac.constants.SystemPermission
 import com.lovelycatv.crystalframework.shared.constants.GlobalConstants
+import com.lovelycatv.crystalframework.shared.exception.ForbiddenException
+import com.lovelycatv.crystalframework.shared.exception.UnauthorizedException
 import com.lovelycatv.crystalframework.shared.response.ApiResponse
 import com.lovelycatv.crystalframework.shared.types.UserAuthentication
+import com.lovelycatv.crystalframework.shared.utils.RbacUtils
 import com.lovelycatv.crystalframework.shared.utils.awaitListWithTimeout
+import com.lovelycatv.crystalframework.tenant.constants.TenantPermission
 import com.lovelycatv.crystalframework.tenant.controller.manager.role.dto.ManagerCreateTenantRoleDTO
 import com.lovelycatv.crystalframework.tenant.controller.manager.role.dto.ManagerDeleteTenantRoleDTO
 import com.lovelycatv.crystalframework.tenant.controller.manager.role.dto.ManagerReadTenantRoleDTO
 import com.lovelycatv.crystalframework.tenant.controller.manager.role.dto.ManagerUpdateTenantRoleDTO
 import com.lovelycatv.crystalframework.tenant.service.manager.TenantRoleManagerService
 import jakarta.validation.Valid
-import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
 
@@ -21,17 +24,25 @@ import org.springframework.web.bind.annotation.*
 class ManagerTenantRoleController(
     private val tenantRoleManagerService: TenantRoleManagerService
 ) {
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_ROLE_READ}')")
     @GetMapping("/list", version = "1")
     suspend fun readAll(
         userAuthentication: UserAuthentication,
         @RequestParam
         tenantId: Long,
     ): ApiResponse<*> {
-        return ApiResponse.success(tenantRoleManagerService.getRepository().findAllByTenantId(tenantId).awaitListWithTimeout())
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_ROLE_READ)) {
+            return ApiResponse.success(tenantRoleManagerService.getRepository().findAllByTenantId(tenantId).awaitListWithTimeout())
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_ROLE_READ_PEM)) {
+            if (tenantId == userAuthentication.tenantId) {
+                return ApiResponse.success(tenantRoleManagerService.getRepository().findAllByTenantId(tenantId).awaitListWithTimeout())
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_ROLE_CREATE}')")
     @PostMapping("/create", version = "1")
     suspend fun create(
         userAuthentication: UserAuthentication,
@@ -39,11 +50,21 @@ class ManagerTenantRoleController(
         @Valid
         dto: ManagerCreateTenantRoleDTO
     ): ApiResponse<*> {
-        tenantRoleManagerService.create(dto)
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_ROLE_CREATE)) {
+            tenantRoleManagerService.create(dto)
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_ROLE_CREATE_PEM)) {
+            userAuthentication.assertTenantIdNotNull()
+            if (dto.tenantId == userAuthentication.tenantId) {
+                tenantRoleManagerService.create(dto)
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
         return ApiResponse.success(null)
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_ROLE_READ}')")
     @GetMapping("/query", version = "1")
     suspend fun query(
         userAuthentication: UserAuthentication,
@@ -51,10 +72,19 @@ class ManagerTenantRoleController(
         @Valid
         dto: ManagerReadTenantRoleDTO
     ): ApiResponse<*> {
-        return ApiResponse.success(tenantRoleManagerService.query(dto))
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_ROLE_READ)) {
+            return ApiResponse.success(tenantRoleManagerService.query(dto))
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_ROLE_READ_PEM)) {
+            if (dto.tenantId == userAuthentication.tenantId) {
+                return ApiResponse.success(tenantRoleManagerService.query(dto))
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_ROLE_UPDATE}')")
     @PostMapping("/update", version = "1")
     suspend fun update(
         userAuthentication: UserAuthentication,
@@ -62,11 +92,21 @@ class ManagerTenantRoleController(
         @Valid
         dto: ManagerUpdateTenantRoleDTO
     ): ApiResponse<*> {
-        tenantRoleManagerService.update(dto)
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_ROLE_UPDATE)) {
+            tenantRoleManagerService.update(dto)
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_ROLE_UPDATE_PEM)) {
+            userAuthentication.assertTenantIdNotNull()
+            if (tenantRoleManagerService.checkIsRelated(dto.id, userAuthentication.tenantId!!)) {
+                tenantRoleManagerService.update(dto)
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
         return ApiResponse.success(null)
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_ROLE_DELETE}')")
     @PostMapping("/delete", version = "1")
     suspend fun delete(
         userAuthentication: UserAuthentication,
@@ -74,7 +114,18 @@ class ManagerTenantRoleController(
         @Valid
         dto: ManagerDeleteTenantRoleDTO
     ): ApiResponse<*> {
-        tenantRoleManagerService.deleteByDTO(dto)
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_ROLE_DELETE)) {
+            tenantRoleManagerService.deleteByDTO(dto)
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_ROLE_DELETE_PEM)) {
+            userAuthentication.assertTenantIdNotNull()
+            if (tenantRoleManagerService.checkIsRelated(dto.ids, userAuthentication.tenantId!!)) {
+                tenantRoleManagerService.deleteByDTO(dto)
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
         return ApiResponse.success(null)
     }
 }
